Only use the 'station' interface for post connections, and set timeouts.
Only start the post thread if the 'station' interface is configured,
and bind the connection to this interface to use it even if there is
also an AP interface.

Set connection timeouts of 60 seconds.

Bump up the post thread stack size a little, as it was overflowing in
current builds.