Rework the enable/disable logging.
The setting saved in the flash parameters now sets the startup state of
logging to flash. Disabling this setting in the config interface no longer
changes the current state of logging. The home page enable/disable button now
starts and pauses the current state of logging, without changing the saved
configuration. This change allows logging to be paused but set to restart
logging when the device restarts.