Fix encoding of the client utimes.
It was not re-calculating the delta on each retry, leading to failure of the
delta encoding.

The logic used to avoid logging frequently repeated utime events was also
reworked to not omit them at the start of a segment - good to have a timestamp
as early as possible in a new segment.

Other uses were updated to follow a similar code pattern.