Rework the walking of the sectors to handle missing indexes.
An index can be missed in the sequence, due to persistent flash write failures
which have been seen and might be due to poor power such as a low battery
etc. The interface to search for a index now returns the new in the sequence
so the caller can advance quickly without retrying intermediate indexes.

The get_buffer_range function has been corrected to give back the flash
semaphor on failure.

The web interface to get a buffer has been corrected and modified. It was
omitting some of the content at the end in some cases, and has been simplified
to just return the data and not the rtc count as a prefix.

The task priority of the data reading and flash writing tasks has been
increased. Logging data and recording it are given priority.